obstacles_detection: fixed ground filtered if under 0

+++ src/nodelets/obstacles_detection.cpp
@@ -167,7 +167,8 @@ private:
 			originalCloud = rtabmap::util3d::transformPointCloud(originalCloud, localTransform);
 			if(maxObstaclesHeight_ > 0)
 			{
-				originalCloud = rtabmap::util3d::passThrough(originalCloud, "z", std::numeric_limits<float>::min(), maxObstaclesHeight_);
+				// std::numeric_limits<float>::lowest() exists only for c++11
+				originalCloud = rtabmap::util3d::passThrough(originalCloud, "z", std::numeric_limits<int>::min(), maxObstaclesHeight_);
 			}
 
 			if(originalCloud->size())
